+++ src/pages/EditTagPage.tsx
@@ -1,0 +1,37 @@
+import { TextInput } from "@mantine/core";
+
+const EditTagPage = () => {
+  return (
+    <>
+      <TextInput
+        label="Name"
+        size="xs"
+        value={tagInput.name}
+        onChange={(e) => setTagInput({ ...tagInput, name: e.target.value })}
+      />
+      <ColorPicker
+        format="hex"
+        value={tagInput.colour}
+        onChange={(newColour) =>
+          setTagInput({ ...tagInput, colour: newColour })
+        }
+      />
+      <Group>
+        <Button size="xs" flex="1" variant="outline" component={Link} to={HOME}>
+          Cancel
+        </Button>
+        <Button
+          size="xs"
+          flex="1"
+          component={Link}
+          to={HOME}
+          onClick={() => createTag(tagInput)}
+        >
+          Save
+        </Button>
+      </Group>
+    </>
+  );
+};
+
+export default EditTagPage;