+++ src/routes/routes.ts
@@ -2,4 +2,5 @@ export const ROOT = "Advanced-Todo";
 
 export const HOME = `/${ROOT}/`;
 export const CREATE_TAG = `/${ROOT}/tag/`;
-export const UPDATE_TODO = `/${ROOT}/:id/`;
+export const EDIT_TODO = `/${ROOT}/:id/`;
+export const EDIT_TAGS = `/${ROOT}/tags/`;